ChronoAnalyser Feature Initialized
- Added plotly.js, flatpickr, js-yaml to project dependencies.
