+++ package.json
@@ -28,8 +28,10 @@
         "@types/color": "^4.2.0",
         "@types/dav": "^1.7.7",
         "@types/jest": "^30.0.0",
+        "@types/js-yaml": "^4.0.9",
         "@types/luxon": "^3.6.2",
         "@types/node": "^24.0.14",
+        "@types/plotly.js": "^3.0.2",
         "@types/react": "^19.1.8",
         "@types/react-dom": "^19.1.6",
         "@typescript-eslint/eslint-plugin": "^8.37.0",
@@ -62,11 +64,14 @@
         "dav": "^1.8.0",
         "deep-equal": "^2.2.3",
         "esbuild-css-modules-plugin": "^3.1.5",
+        "flatpickr": "^4.6.13",
         "ical": "^0.8.0",
         "ical.js": "^1.5.0",
+        "js-yaml": "^4.1.0",
         "luxon": "^3.7.1",
         "moment": "^2.30.1",
         "obsidian-daily-notes-interface": "^0.9.4",
+        "plotly.js": "^3.0.1",
         "react": "^19.1.0",
         "react-dom": "^19.1.0",
         "rrule": "^2.8.1",
